refactor(flight-tune): split terminal integration boundaries (#457)
