+++ tools/flight-tune/src/journal/replay.rs
@@ -1,12 +1,13 @@
 use crate::journal::{
     AttemptRole, CampaignPhase, FinalQualificationOutcome, JOURNAL_SCHEMA_VERSION, JournalEntry,
-    JournalEvent, OperationStatus, PromotionDecision, SessionIdentity,
+    JournalEvent, PromotionDecision, SessionIdentity,
 };
 use crate::{
     CandidateEvaluation, CandidateTransitionReference, Digest, SearchStage, TrainingObservation,
     TuneError,
 };
 
+mod attempt;
 mod plan;
 mod run;
 pub(crate) mod transition;
@@ -160,7 +161,7 @@ fn apply_event(
             candidate,
             plan_digest,
             transition,
-        } => prepare(
+        } => attempt::prepare(
             state,
             *trial_id,
             *role,
@@ -178,7 +179,7 @@ fn apply_event(
             trial_id,
             evaluation,
             selected_as_training_incumbent,
-        } => complete(
+        } => attempt::complete(
             state,
             *trial_id,
             evaluation,
@@ -187,13 +188,13 @@ fn apply_event(
             fixed_seed,
         ),
         JournalEvent::AttemptQuarantined { trial_id, reason } => {
-            quarantine(state, *trial_id, reason)
+            attempt::quarantine(state, *trial_id, reason)
         }
         JournalEvent::CleanupRecorded {
             trial_id,
             stop,
             cleanup: cleanup_status,
-        } => cleanup(state, *trial_id, stop, cleanup_status),
+        } => attempt::cleanup(state, *trial_id, stop, cleanup_status),
         JournalEvent::Frozen {
             baseline,
             candidate,
@@ -203,279 +204,6 @@ fn apply_event(
             close::seal(state, *candidate, outcome, initial, stage)
         }
     }
-}
-
-#[allow(clippy::too_many_arguments)]
-fn prepare(
-    state: &mut JournalState,
-    trial_id: u64,
-    role: AttemptRole,
-    candidate: Digest,
-    plan_digest: Digest,
-    transition_reference: Option<&CandidateTransitionReference>,
-    stage: &SearchStage,
-    initial: Digest,
-    fixed_seed: u64,
-) -> Result<(), TuneError> {
-    let expected_plan = role.plan_digest(stage, candidate, fixed_seed)?;
-    if state.pending.is_some()
-        || trial_id != state.next_trial_id
-        || candidate.is_zero()
-        || plan_digest != expected_plan
-        || !role_allowed(state, role, candidate, initial)
-    {
-        return Err(invalid(
-            "an attempt preparation has invalid state or identity",
-        ));
-    }
-    transition::validate_attempt(state, role, candidate, transition_reference)?;
-    state.pending = Some(PendingAttempt {
-        trial_id,
-        role,
-        candidate,
-        transition: transition_reference.cloned(),
-        prepared_runs: Vec::new(),
-        outcome: None,
-    });
-    state.authorized_transition = None;
-    state.next_trial_id = state.next_trial_id.wrapping_add(1);
-    Ok(())
-}
-
-fn role_allowed(
-    state: &JournalState,
-    role: AttemptRole,
-    candidate: Digest,
-    initial: Digest,
-) -> bool {
-    match role {
-        AttemptRole::TrainingBaseline => {
-            state.phase == CampaignPhase::Searching
-                && state.training_baseline.is_none()
-                && candidate == initial
-        }
-        AttemptRole::TrainingChallenger { attempt_index } => {
-            state.phase == CampaignPhase::Searching
-                && has_passed_training_baseline_and_incumbent(state)
-                && attempt_index == state.training_attempt_count
-        }
-        AttemptRole::PromotionBaseline => {
-            state.phase == CampaignPhase::Frozen
-                && state.promotion_baseline.is_none()
-                && candidate == initial
-        }
-        AttemptRole::PromotionFrozen => {
-            state.phase == CampaignPhase::Frozen
-                && state.promotion_baseline.is_some()
-                && state.promotion_frozen.is_none()
-                && state.frozen_candidate == Some(candidate)
-        }
-        AttemptRole::FinalQualification => {
-            state.phase == CampaignPhase::PromotionClosed
-                && state.final_evaluation.is_none()
-                && candidate == state.selected_release_candidate(initial)
-        }
-    }
-}
-
-fn has_passed_training_baseline_and_incumbent(state: &JournalState) -> bool {
-    matches!(
-        &state.training_baseline,
-        Some(CandidateEvaluation::Passed { .. })
-    ) && matches!(
-        &state.training_incumbent_evaluation,
-        Some(CandidateEvaluation::Passed { .. })
-    )
-}
-
-fn complete(
-    state: &mut JournalState,
-    trial_id: u64,
-    evaluation: &CandidateEvaluation,
-    selected: Option<bool>,
-    stage: &SearchStage,
-    fixed_seed: u64,
-) -> Result<(), TuneError> {
-    let role = state
-        .pending
-        .as_ref()
-        .filter(|pending| pending.trial_id == trial_id && pending.outcome.is_none())
-        .map(|pending| pending.role)
-        .ok_or_else(|| invalid("the attempt is not pending or already has an outcome"))?;
-    plan::validate_evaluation(evaluation, role, stage, fixed_seed)?;
-    let pending = state
-        .pending
-        .as_ref()
-        .ok_or_else(|| invalid("the attempt lost its run preparation"))?;
-    run::validate_outcome(pending, evaluation)?;
-    validate_training_selection(state, role, evaluation, selected)?;
-    let pending = pending_without_outcome(state, trial_id)?;
-    pending.outcome = Some(PendingOutcome {
-        evaluation: evaluation.clone(),
-        selected,
-    });
-    Ok(())
-}
-
-fn validate_training_selection(
-    state: &JournalState,
-    role: AttemptRole,
-    evaluation: &CandidateEvaluation,
-    selected: Option<bool>,
-) -> Result<(), TuneError> {
-    let expected = match role {
-        AttemptRole::TrainingBaseline => Some(evaluation.aggregate().is_some()),
-        AttemptRole::TrainingChallenger { .. } => Some(training_better(
-            state.training_incumbent_evaluation.as_ref(),
-            evaluation,
-        )),
-        AttemptRole::PromotionBaseline
-        | AttemptRole::PromotionFrozen
-        | AttemptRole::FinalQualification => None,
-    };
-    if selected != expected {
-        return Err(invalid("a training incumbent decision is not reproducible"));
-    }
-    Ok(())
-}
-
-fn training_better(
-    incumbent: Option<&CandidateEvaluation>,
-    challenger: &CandidateEvaluation,
-) -> bool {
-    let Some(challenger_score) = challenger.aggregate() else {
-        return false;
-    };
-    incumbent
-        .and_then(CandidateEvaluation::aggregate)
-        .is_none_or(|incumbent_score| challenger_score.mean_loss < incumbent_score.mean_loss)
-}
-
-fn quarantine(state: &mut JournalState, trial_id: u64, reason: &str) -> Result<(), TuneError> {
-    if reason.trim().is_empty() || reason.len() > 4_096 {
-        return Err(invalid("a quarantine reason is empty or too long"));
-    }
-    let pending = pending_without_outcome(state, trial_id)?;
-    let selected = match pending.role {
-        AttemptRole::TrainingBaseline | AttemptRole::TrainingChallenger { .. } => Some(false),
-        _ => None,
-    };
-    pending.outcome = Some(PendingOutcome {
-        evaluation: CandidateEvaluation::Quarantined {
-            reason: reason.to_owned(),
-        },
-        selected,
-    });
-    Ok(())
-}
-
-fn cleanup(
-    state: &mut JournalState,
-    trial_id: u64,
-    stop: &OperationStatus,
-    cleanup: &OperationStatus,
-) -> Result<(), TuneError> {
-    validate_operation_status(stop)?;
-    validate_operation_status(cleanup)?;
-    let Some(pending) = state.pending.as_ref() else {
-        return Err(invalid("cleanup has no pending attempt"));
-    };
-    if pending.trial_id != trial_id || pending.outcome.is_none() {
-        return Err(invalid("cleanup has the wrong trial or no saved outcome"));
-    }
-    if cleanup.succeeded() {
-        let completed = state
-            .pending
-            .take()
-            .ok_or_else(|| invalid("cleanup lost its pending attempt"))?;
-        finalize(state, completed)?;
-    }
-    Ok(())
-}
-
-fn finalize(state: &mut JournalState, pending: PendingAttempt) -> Result<(), TuneError> {
-    let outcome = pending
-        .outcome
-        .ok_or_else(|| invalid("a clean attempt has no outcome"))?;
-    match pending.role {
-        AttemptRole::TrainingBaseline => finalize_baseline(state, outcome),
-        AttemptRole::TrainingChallenger { attempt_index } => {
-            finalize_training(state, pending.candidate, attempt_index, outcome)
-        }
-        AttemptRole::PromotionBaseline => {
-            state.promotion_baseline = Some(outcome.evaluation);
-            Ok(())
-        }
-        AttemptRole::PromotionFrozen => {
-            state.promotion_frozen = Some(outcome.evaluation);
-            Ok(())
-        }
-        AttemptRole::FinalQualification => {
-            state.final_evaluation = Some(outcome.evaluation);
-            Ok(())
-        }
-    }
-}
-
-fn finalize_baseline(state: &mut JournalState, outcome: PendingOutcome) -> Result<(), TuneError> {
-    if outcome.selected == Some(true) {
-        state.training_incumbent_evaluation = Some(outcome.evaluation.clone());
-    }
-    state.training_baseline = Some(outcome.evaluation);
-    Ok(())
-}
-
-fn finalize_training(
-    state: &mut JournalState,
-    candidate: Digest,
-    attempt_index: u64,
-    outcome: PendingOutcome,
-) -> Result<(), TuneError> {
-    if attempt_index != state.training_attempt_count {
-        return Err(invalid("a completed training attempt has the wrong index"));
-    }
-    let selected = outcome.selected == Some(true);
-    let observation = TrainingObservation {
-        attempt_index,
-        candidate_digest: candidate,
-        selected_as_incumbent: selected,
-        hard_gate_failed: matches!(
-            outcome.evaluation,
-            CandidateEvaluation::HardGateFailed { .. }
-        ),
-        quarantined: matches!(outcome.evaluation, CandidateEvaluation::Quarantined { .. }),
-        training_mean_loss: outcome
-            .evaluation
-            .aggregate()
-            .map(|aggregate| aggregate.mean_loss),
-    };
-    if selected {
-        state.training_incumbent = candidate;
-        state.training_incumbent_evaluation = Some(outcome.evaluation.clone());
-    }
-    state.training_history.push(observation);
-    state.training_attempt_count = state.training_attempt_count.wrapping_add(1);
-    Ok(())
-}
-
-fn pending_without_outcome(
-    state: &mut JournalState,
-    trial_id: u64,
-) -> Result<&mut PendingAttempt, TuneError> {
-    state
-        .pending
-        .as_mut()
-        .filter(|pending| pending.trial_id == trial_id && pending.outcome.is_none())
-        .ok_or_else(|| invalid("the attempt is not pending or already has an outcome"))
-}
-
-fn validate_operation_status(status: &OperationStatus) -> Result<(), TuneError> {
-    if let OperationStatus::Failed { detail } = status
-        && detail.trim().is_empty()
-    {
-        return Err(invalid("a cleanup failure detail is empty"));
-    }
-    Ok(())
 }
 
 fn invalid(detail: impl Into<String>) -> TuneError {
